Rename file to use underscores; drop unnecessary zeroing
